rename armour to wearable
